refactor: land U.11 atm-graft cleanup
